ai-engine/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
gunicorn==21.2.0
scikit-learn>=1.3.0
numpy>=1.26.0
python-dotenv==1.0.0
google-generativeai==0.3.2
